history: surface recording and move tests
