+++ src/commands/dedup.rs
@@ -704,7 +704,7 @@ type SourcesByRev<'a> = BTreeMap<&'a str, BTreeMap<&'a str, Vec<Vec<String>>>>;
 fn group_sources_by_rev(entries: &[Entry]) -> SourcesByRev<'_> {
     let mut by_rev = BTreeMap::<&str, BTreeMap<&str, Vec<Vec<String>>>>::new();
     for entry in entries {
-        let names = &mut by_rev.entry(entry_compare_rev(entry)).or_default();
+        let names = by_rev.entry(entry_compare_rev(entry)).or_default();
         names
             .entry(entry.name.as_str())
             .or_default()
@@ -798,4 +798,309 @@ pub(super) fn pick_name(id: &SourceId, aliases: &BTreeSet<String>) -> String {
         .min_by_key(|name| (name.len(), name.as_str()))
         .cloned()
         .unwrap_or_default()
+}
+
+#[cfg(test)]
+mod tests {
+    use std::{
+        collections::{
+            BTreeMap,
+            BTreeSet,
+            HashMap,
+        },
+        iter,
+    };
+
+    use super::{
+        Entry,
+        MAX_COMPARE_JOBS,
+        Side,
+        apply_follows,
+        classify,
+        comparator,
+        compare_jobs,
+        group_diverges,
+        pick_name,
+        rev_last_modified,
+    };
+    use crate::{
+        fetch::github::CompareStatus,
+        report::Mark,
+        source::id::SourceId,
+    };
+
+    fn map(pairs: &[(&str, &str)]) -> BTreeMap<String, String> {
+        pairs
+            .iter()
+            .map(|&(alias, target)| (alias.to_owned(), target.to_owned()))
+            .collect()
+    }
+
+    fn set(items: &[&str]) -> BTreeSet<String> {
+        items.iter().map(|item| (*item).to_owned()).collect()
+    }
+
+    fn tack_entry(path: &[&str], name: &str, rev: &str, lm: Option<u64>) -> Entry {
+        Entry {
+            side: Side::Tack,
+            ..entry(path, name, rev, lm)
+        }
+    }
+
+    fn entry(path: &[&str], name: &str, rev: &str, lm: Option<u64>) -> Entry {
+        entry_full(path, name, rev, rev, lm)
+    }
+
+    fn entry_full(
+        path: &[&str],
+        name: &str,
+        _display_rev: &str,
+        rev: &str,
+        lm: Option<u64>,
+    ) -> Entry {
+        Entry {
+            path: path.iter().map(|item| (*item).to_owned()).collect(),
+            name: name.to_owned(),
+            side: Side::Flake,
+            rev: rev.to_owned(),
+            lm,
+        }
+    }
+
+    fn source_id(str: &str) -> SourceId {
+        SourceId::from_url(str).unwrap()
+    }
+
+    #[test]
+    fn pick_name_strips_dot_nix_and_flattens_dots() {
+        assert_eq!(
+            pick_name(
+                &source_id("github:cachix/git-hooks.nix"),
+                &set(&["git-hooks"])
+            ),
+            "git-hooks"
+        );
+        assert_eq!(
+            pick_name(
+                &source_id("github:nix-community/nixpkgs.lib"),
+                &set(&["nixpkgs-lib"])
+            ),
+            "nixpkgs-lib"
+        );
+    }
+
+    #[test]
+    fn pick_name_falls_back_to_shortest_alias_for_non_github() {
+        let aliases = set(&["my-pin", "the-tarball"]);
+        assert_eq!(pick_name(&source_id("https://x/y"), &aliases), "my-pin");
+    }
+
+    #[test]
+    fn comparator_prefers_top_level_even_without_last_modified() {
+        let entries = vec![
+            entry(&["parent"], "aaa", "newer", Some(20)),
+            entry(&[], "top", "top-rev", None),
+        ];
+        assert_eq!(
+            comparator(&entries).map(|entry| (entry.rev.as_str(), entry.lm)),
+            Some(("top-rev", None))
+        );
+    }
+
+    #[test]
+    fn comparator_uses_newest_known_transitive_then_deterministic_fallback() {
+        let entries_with_known_time = vec![
+            entry(&["parent"], "aaa", "unknown", None),
+            entry(&["parent"], "bbb", "older", Some(10)),
+            entry(&["parent"], "ccc", "newer", Some(20)),
+        ];
+        assert_eq!(
+            comparator(&entries_with_known_time).map(|entry| (entry.rev.as_str(), entry.lm)),
+            Some(("newer", Some(20)))
+        );
+
+        let entries_without_times = vec![
+            entry(&["parent"], "bbb", "unknown-b", None),
+            entry(&["parent"], "aaa", "unknown-a", None),
+        ];
+        assert_eq!(
+            comparator(&entries_without_times).map(|entry| (entry.rev.as_str(), entry.lm)),
+            Some(("unknown-a", None))
+        );
+    }
+
+    #[test]
+    fn group_divergence_uses_semantic_revs() {
+        let entries = vec![
+            entry_full(
+                &[],
+                "base",
+                "abcdef0",
+                "abcdef0000000000000000000000000000000000",
+                Some(10),
+            ),
+            entry_full(
+                &["dep"],
+                "head",
+                "abcdef0",
+                "abcdef0999999999999999999999999999999999",
+                Some(20),
+            ),
+        ];
+
+        assert!(group_diverges(&entries));
+    }
+
+    #[test]
+    fn compare_jobs_use_semantic_revs() {
+        let mut groups = BTreeMap::new();
+        groups.insert(source_id("github:o/r"), vec![
+            entry_full(
+                &[],
+                "base",
+                "1111111",
+                "1111111111111111111111111111111111111111",
+                Some(10),
+            ),
+            entry_full(
+                &["dep"],
+                "head",
+                "2222222",
+                "2222222222222222222222222222222222222222",
+                Some(20),
+            ),
+        ]);
+
+        let (jobs, capped) = compare_jobs(&groups);
+
+        assert_eq!(capped, 0);
+        assert_eq!(jobs.len(), 1);
+        assert_eq!(jobs[0].base, "1111111111111111111111111111111111111111");
+        assert_eq!(jobs[0].head, "2222222222222222222222222222222222222222");
+    }
+
+    #[test]
+    fn compare_jobs_are_capped_before_network_work() {
+        let mut entries = vec![entry_full(&[], "base", "base", "base-full", Some(0))];
+        for i in 0..(MAX_COMPARE_JOBS + 5) {
+            entries.push(entry_full(
+                &["dep"],
+                &format!("head-{i:03}"),
+                &format!("h{i:06}"),
+                &format!("head-full-{i:03}"),
+                Some(u64::try_from(i).unwrap() + 1),
+            ));
+        }
+        let mut groups = BTreeMap::new();
+        groups.insert(source_id("github:o/r"), entries);
+
+        let (jobs, capped) = compare_jobs(&groups);
+
+        assert_eq!(jobs.len(), MAX_COMPARE_JOBS);
+        assert_eq!(capped, 5);
+    }
+
+    #[test]
+    fn classify_prefers_branch_status_over_misleading_timestamps() {
+        let id = source_id("github:o/r");
+        let entries = vec![
+            entry(&[], "base", "base", Some(500)),
+            entry(&["dep"], "head", "head", Some(100)),
+        ];
+        let compares = HashMap::from([((id.clone(), "head".to_owned()), CompareStatus::Ahead)]);
+
+        let mark = classify(
+            &id,
+            "head",
+            comparator(&entries),
+            &rev_last_modified(&entries),
+            &compares,
+        );
+
+        assert_eq!(mark, Mark::Ahead);
+    }
+
+    #[test]
+    fn classify_reports_diverged_branch_status() {
+        let id = source_id("github:o/r");
+        let entries = vec![
+            entry(&[], "base", "base", Some(100)),
+            entry(&["dep"], "head", "head", Some(200)),
+        ];
+        let compares = HashMap::from([((id.clone(), "head".to_owned()), CompareStatus::Diverged)]);
+
+        let mark = classify(
+            &id,
+            "head",
+            comparator(&entries),
+            &rev_last_modified(&entries),
+            &compares,
+        );
+
+        assert_eq!(mark, Mark::Diverged);
+    }
+
+    #[test]
+    fn classify_distinguishes_timestamp_fallback() {
+        let id = source_id("github:o/r");
+        let entries = vec![
+            entry(&[], "base", "base", Some(100)),
+            entry(&["dep"], "head", "head", Some(200)),
+        ];
+        let compares = HashMap::<(SourceId, String), CompareStatus>::new();
+
+        let mark = classify(
+            &id,
+            "head",
+            comparator(&entries),
+            &rev_last_modified(&entries),
+            &compares,
+        );
+
+        assert_eq!(mark, Mark::DatedNewer);
+    }
+
+    #[test]
+    fn apply_follows_syncs_rev_and_lm_to_target() {
+        let id = source_id("github:o/r");
+        let mut groups = BTreeMap::new();
+        groups.insert(id.clone(), vec![
+            entry(&[], "nixpkgs", "newrev-full", Some(100)),
+            entry(&["dep"], "nixpkgs-lib", "oldrev", Some(50)),
+        ]);
+        let by_name = BTreeMap::new();
+        let all_follow = map(&[("nixpkgs-lib", "nixpkgs")]);
+        let top_revs = map(&[("nixpkgs", "newrev-full")]);
+        let top_lms = iter::once(("nixpkgs".to_owned(), 100_u64)).collect();
+
+        apply_follows(&mut groups, &by_name, &all_follow, &top_revs, &top_lms);
+
+        let followed = &groups[&id][1];
+        assert_eq!(followed.rev, "newrev-full");
+        assert_eq!(followed.lm, Some(100));
+    }
+
+    #[test]
+    fn apply_follows_honors_scoped_all_follow_per_side() {
+        let id = source_id("github:o/r");
+        let mut groups = BTreeMap::new();
+        groups.insert(id.clone(), vec![tack_entry(
+            &["parent"],
+            "dep",
+            "oldrev",
+            Some(50),
+        )]);
+        let by_name = BTreeMap::new();
+        let top_revs = map(&[("replacement", "newrev-full")]);
+        let top_lms = iter::once(("replacement".to_owned(), 100_u64)).collect();
+
+        let flake_rule = map(&[("flake:dep", "replacement")]);
+        apply_follows(&mut groups, &by_name, &flake_rule, &top_revs, &top_lms);
+        assert_eq!(groups[&id][0].rev, "oldrev");
+
+        let tack_rule = map(&[("tack:dep", "replacement")]);
+        apply_follows(&mut groups, &by_name, &tack_rule, &top_revs, &top_lms);
+        assert_eq!(groups[&id][0].rev, "newrev-full");
+        assert_eq!(groups[&id][0].lm, Some(100));
+    }
 }
