+++ src/commands/edit.rs
@@ -120,3 +120,41 @@ pub fn alias(project: &Project, name: &str, template: Option<&str>, remove: bool
     }
     Ok(())
 }
+
+#[cfg(test)]
+mod tests {
+    use std::fs;
+
+    use super::rm_in_dir;
+
+    #[test]
+    fn rm_removes_orphaned_lock_entry() {
+        let dir = tempfile::tempdir().unwrap();
+        fs::write(dir.path().join("pins.toml"), "[inputs]\n").unwrap();
+        fs::write(
+            dir.path().join("pins.lock.json"),
+            r#"{"gone":{"type":"github","owner":"o","repo":"r","rev":"bad","narHash":"sha256-x"}}"#,
+        )
+        .unwrap();
+
+        assert_eq!(rm_in_dir(dir.path(), "gone").unwrap(), (false, true));
+        assert_eq!(
+            fs::read_to_string(dir.path().join("pins.toml")).unwrap(),
+            "[inputs]\n"
+        );
+        assert_eq!(
+            fs::read_to_string(dir.path().join("pins.lock.json")).unwrap(),
+            "{}\n"
+        );
+    }
+
+    #[test]
+    fn rm_errors_when_pin_and_lock_are_missing() {
+        let dir = tempfile::tempdir().unwrap();
+        fs::write(dir.path().join("pins.toml"), "[inputs]\n").unwrap();
+        fs::write(dir.path().join("pins.lock.json"), "{}\n").unwrap();
+
+        let err = rm_in_dir(dir.path(), "missing").unwrap_err().to_string();
+        assert_eq!(err, "no input 'missing'");
+    }
+}
